Key 'feeds' table off a numeric primary key (breaking DB change)
The previous approach, of using the feed URL as the primary key, was idiotic, and led to bugs.  Use this change also as a refactoring exercise, making fetch.js much cleaner.

Because SQLite can only add columns and not remove them, there was no way to make a migration for this change.  I think it would be good to switch to using a DB server soon.