Event edit form: make end date default to start date if not set
To do this I had to upgrade to the latest datetimepicker, which led me to realise that 'bower install', unlike 'npm install', doesn't update to the latest version of packages, so I replaced it in 'npm run deploy' with bower update.  Additionally, update bower.json to allow automatic updates to later versions of packages.