Add newsletter sending functionality from web UI
This commit adds a newsletter sending UI at /admin/newsletter which
gives control over which events to include (chosen by date range) and
allows a message, written using Markdown, to be put at the start of the
email.  It also uses a fancier email template.

This has required minor changes to the JSON API (it returns an extra
URL field).  It is currently not linked to from anywhere else in the
site.